added two separate files for edit profile based on user
