+++ src/App.jsx
@@ -8,12 +8,14 @@ import LoginPage from './Pages/LoginPage.jsx';
 import FoodPantryDashboard from './Pages/FoodPantryDashboard';
 import RestaurantDashboard from './Pages/RestaurantDashboard';
 import AddListingPage from './Pages/AddListingPage';
-import EditProfilePage from './Pages/EditProfilePage'
+
 import CartPage from './Pages/CartPage.jsx';
 import DonationHistoryPage from './Pages/DonationHistoryPage.jsx';
 import AllRestaurantsPage from './Pages/AllRestaurantsPage.jsx';
 import FoodPantryProfilePage from './Pages/FoodPantryProfilePage.jsx';
 import RestaurantProfilePage from './Pages/RestaurantProfilePage.jsx';
+import EditProfilePageRestaurant from './Pages/EditProfilePageRestaurant'
+import EditProfilePageFoodPantry from './Pages/EditProfilePageFoodPantry'
 
 
 
@@ -28,7 +30,8 @@ const App = () => {
           <Route path="/sign-up" element={<SignUpPage />} />
           <Route path="/food-dashboard/:username" element={<FoodPantryDashboard />} />
           <Route path="/restaurant-dashboard/:username" element={<RestaurantDashboard />} />
-          <Route path="/edit-profile" element={<EditProfilePage />} />
+          <Route path="/edit-profile-restaurant" element={<EditProfilePageRestaurant />} />
+          <Route path="/edit-profile-food" element={<EditProfilePageFoodPantry />} />
           <Route path="/add-listing" element={<AddListingPage />} />
           <Route path="/cart-page" element={<CartPage />} />
           <Route path="/donation-history" element={<DonationHistoryPage />} />
