+++ src/Components/DashboardHeader/DashboardHeader.css
@@ -2,7 +2,7 @@
     display: flex;
     justify-content: space-between;
     align-items: center;
-    width: 250%;
+    width: 100%;
     padding: 20px;
     border-bottom: 2px solid #000;
     color: #004715;
