+++ src/Components/FoodAvailableHeader/FoodAvailableHeader.css
@@ -2,7 +2,7 @@
     display: flex;
     justify-content: space-between;
     align-items: center;
-    width: 250%;
+    width: 100%;
     padding-top: 10px;
 }
