+++ src/Pages/LandingPage.css
@@ -12,7 +12,7 @@
 }
 
 .impact {
-    background-color: #004715;
+    /* background-color: #004715; */
     display: flex;
     flex-direction: row;
     flex-wrap: wrap;
